mail file source code
This is a Java program that maintains a list of names and phone numbers,
or a phone directory of your friends. For this program, the class DirectoryEntry is used to represent each
item in the phone directory. The class PhoneDirectory has the ArrayList<DirectoryEntry> theDirectory as
its private data field and the following methods (in addition to its constructor and accessor methods)